+++ ohouse/src/main/java/com/ohouse/ohouse/service/UserService.java
@@ -6,6 +6,6 @@ public interface UserService {
 
   UserDTO getUserByEmail(String email);
 
-  void savePhoneNumberAndMarkVerified(Long userId, String phoneNumber);
+  void savePhoneNumberAndMarkVerified(int userId, String phoneNumber);
 
 }
